requirements.txt
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==2.0.12
colorama==0.4.6
exceptiongroup==1.2.2
flake8==5.0.4
flake8-docstrings==1.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.1
pluggy==1.5.0
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
pyTelegramBotAPI==4.14.1
pytest==7.1.3
pytest-timeout==2.1.0
python-dotenv==0.20.0
python-telegram-bot==21.6
requests==2.26.0
sniffio==1.3.1
snowballstemmer==2.2.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.20
